example/devsecops-pipeline/sample-app/requirements.txt
# Sample app dependencies (some with known vulnerabilities for SCA testing)
flask==2.3.0
pyyaml==6.0
requests==2.28.0
jinja2==3.1.2
